+++ src/components/pages/StudentsPage/Chatbox.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 
 import { Box } from '@mui/material';
-import { useEffect, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
 import { scrollToBottomOfElement } from '@utils/classrooms';
 import chatboxCSS from './Chatbox.css';
@@ -10,6 +10,40 @@ import SendMessages from './SendMessages';
 import CopyButton from '@components/shared/CopyButton';
 
 export default function Chatbox({ socket, chat, setChat, chatEndedMsg }) {
+  const [peerIsTyping, setPeerIsTyping] = useState(false);
+
+  function addChatMessage(sender, message) {
+    setChat((chat) => ({
+      ...chat,
+      conversation: [...chat.conversation, [sender, message]],
+    }));
+  }
+
+  useEffect(() => {
+    if (socket) {
+      socket.on('student sent message', ({ message }) => {
+        setPeerIsTyping(false);
+        addChatMessage('peer', message);
+      });
+
+      socket.on('teacher sent message', ({ message }) => {
+        addChatMessage('teacher', message);
+      });
+
+      socket.on('solo mode: teacher sent message', ({ message }) => {
+        addChatMessage('teacher', message);
+      });
+    }
+
+    return () => {
+      if (socket) {
+        socket.off('student sent message');
+        socket.off('teacher sent message');
+        socket.off('solo mode: teacher sent message');
+      }
+    };
+  }, [setChat, socket]);
+
   useEffect(() => {
     scrollToBottomOfElement(chatboxConversationContainer);
   }, [chat.conversation]);
@@ -20,7 +54,7 @@ export default function Chatbox({ socket, chat, setChat, chatEndedMsg }) {
     <Box css={chatboxCSS.chatboxContainer}>
       <CopyButton elementId='conversation' />
       <Box css={chatboxCSS.chatboxTop} ref={chatboxConversationContainer}>
-        <Conversation socket={socket} chat={chat} setChat={setChat} />
+        <Conversation chat={chat} />
       </Box>
       <Box css={chatboxCSS.chatboxBottom}>
         <SendMessages
@@ -28,6 +62,8 @@ export default function Chatbox({ socket, chat, setChat, chatEndedMsg }) {
           chat={chat}
           setChat={setChat}
           chatEndedMsg={chatEndedMsg}
+          peerIsTyping={peerIsTyping}
+          setPeerIsTyping={setPeerIsTyping}
         />
       </Box>
     </Box>
